refactor: remove avilla.core.abstract
