+++ avilla/core/dispatchers.py
@@ -6,8 +6,8 @@ from typing import TYPE_CHECKING
 from graia.broadcast.entities.dispatcher import BaseDispatcher
 
 from avilla.core._runtime import ctx_protocol
-from avilla.core.abstract.account import AbstractAccount
-from avilla.core.abstract.event import AvillaEvent
+from avilla.core.account import AbstractAccount
+from avilla.core.event import AvillaEvent
 from avilla.core.context import Context
 from avilla.core.protocol import BaseProtocol
