+++ avilla/core/elements.py
@@ -6,7 +6,7 @@ from typing import Any
 from graia.amnesia.message import Element
 from graia.amnesia.message.element import Text as Text
 
-from avilla.core.abstract.resource import LocalFileResource, Resource
+from avilla.core.resource import LocalFileResource, Resource
 from avilla.core.utilles.selector import Selector
