+++ avilla/core/message.py
@@ -8,7 +8,7 @@ from graia.amnesia.message import __message_chain_class__
 from avilla.core.platform import Land
 from avilla.core.utilles.selector import Selector
 
-from ...spec.core.message.skeleton import MessageRevoke
+from ..spec.core.message.skeleton import MessageRevoke
 from .metadata import Metadata
 
 if TYPE_CHECKING:
