+++ avilla/spec/core/privilege/skeleton.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import timedelta
 
-from avilla.core.abstract.trait import Fn, Trait
+from avilla.core.trait import Fn, Trait
 from avilla.core.utilles.selector import Selector
