+++ avilla/spec/core/profile/metadata.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from avilla.core.abstract.metadata import Metadata
+from avilla.core.metadata import Metadata
 
 
 @dataclass
